Indian-Sign-Language-Gesture-Recognition-master/requirements-updated.txt
# Core Django
Django==3.2.23
django-crispy-forms==1.14.0
python-decouple==3.8

# Machine Learning & Computer Vision
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.1.78
scikit-learn==1.3.2
scikit-image==0.22.0

# Audio Processing (Cross-platform)
SpeechRecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1

# Data Science
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.3
Pillow==10.1.0

# Utilities
h5py==3.10.0
joblib==1.3.2
pyaudio==0.2.14

# Optional: For better performance
# tensorflow-metal==1.0.0  # Uncomment for Mac M1/M2 GPU acceleration
